+++ src/Componentes/NavBar/index.tsx
@@ -1,38 +1,208 @@
-import React from 'react';
-import SearchBar from 'Componentes/SearchBar';
-import { Link } from 'react-router-dom';
-import { setHome } from 'redux/slices/mascotas';
-import { useCustomDispatch } from 'hooks/redux';
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Badge from '@mui/material/Badge';
+import MenuItem from '@mui/material/MenuItem';
+import Menu from '@mui/material/Menu';
+// import MenuIcon from '@mui/icons-material/Menu';
+import AccountCircle from '@mui/icons-material/AccountCircle';
+import MailIcon from '@mui/icons-material/Mail';
+import NotificationsIcon from '@mui/icons-material/Notifications';
+import MoreIcon from '@mui/icons-material/MoreVert';
 import Button from '@mui/material/Button';
+import { Link } from 'react-router-dom';
+import FavoriteIcon from '@mui/icons-material/Favorite';
+import Logo from '../../Assets/image/LOGO.jpg';
+import SearchBar from './searchBar';
 import './styleNavbar.css';
-import { Home } from '@mui/icons-material';
 
-const NavBar: React.FC = () => {
-  const dispatch = useCustomDispatch();
-  const handleSubmit = (e: React.MouseEvent<HTMLAnchorElement>): void => {
-    dispatch(setHome());
+export default function NavBar(): any {
+  const rol = localStorage.getItem('rol');
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const [mobileMoreAnchorEl, setMobileMoreAnchorEl] =
+    React.useState<null | HTMLElement>(null);
+
+  const isMenuOpen = Boolean(anchorEl);
+  const isMobileMenuOpen = Boolean(mobileMoreAnchorEl);
+
+  const handleProfileMenuOpen = (event: React.MouseEvent<HTMLElement>): any => {
+    setAnchorEl(event.currentTarget);
   };
 
-  return (
-    <div className="navbar">
-      <Link to={'/'} style={{ textDecoration: 'none' }} onClick={handleSubmit}>
-        <Button>Home</Button>
-      </Link>
-      <Link to={'/donacion'} style={{ textDecoration: 'none' }}>
-        <Button>Donacion</Button>
-      </Link>
-      <Link to={'/favoritos'} style={{ textDecoration: 'none' }}>
-        <Button>Favoritos</Button>
-      </Link>
-      <Link to={'/publicar'} style={{ textDecoration: 'none' }}>
-        <Button>Publicar Mascota</Button>
-      </Link>
-      <Link to={'/misPublicaciones'} style={{ textDecoration: 'none' }}>
-        <Button>Mis Publicaciones</Button>
-      </Link>
-      <SearchBar />
-    </div>
-  );
-};
+  const handleMobileMenuClose = (): any => {
+    setMobileMoreAnchorEl(null);
+  };
 
-export default NavBar;
+  const handleMenuClose = (): any => {
+    setAnchorEl(null);
+    handleMobileMenuClose();
+  };
+
+  const handleMobileMenuOpen = (event: React.MouseEvent<HTMLElement>): any => {
+    setMobileMoreAnchorEl(event.currentTarget);
+  };
+
+  const menuId = 'primary-search-account-menu';
+  const renderMenu = (
+    <Menu
+      anchorEl={anchorEl}
+      anchorOrigin={{
+        vertical: 'top',
+        horizontal: 'right'
+      }}
+      id={menuId}
+      keepMounted
+      transformOrigin={{
+        vertical: 'top',
+        horizontal: 'right'
+      }}
+      open={isMenuOpen}
+      onClose={handleMenuClose}
+    >
+      <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
+      <MenuItem onClick={handleMenuClose}>My account</MenuItem>
+      {rol === 'administrador' && (
+        <Link to={'/dashboard'}>
+          <MenuItem onClick={handleMenuClose}>Dashboard</MenuItem>
+        </Link>
+      )}
+    </Menu>
+  );
+
+  const mobileMenuId = 'primary-search-account-menu-mobile';
+  const renderMobileMenu = (
+    <Menu
+      anchorEl={mobileMoreAnchorEl}
+      anchorOrigin={{
+        vertical: 'top',
+        horizontal: 'right'
+      }}
+      id={mobileMenuId}
+      keepMounted
+      transformOrigin={{
+        vertical: 'top',
+        horizontal: 'right'
+      }}
+      open={isMobileMenuOpen}
+      onClose={handleMobileMenuClose}
+    >
+      <MenuItem>
+        <IconButton size="large" aria-label="show 4 new mails" color="inherit">
+          <Badge badgeContent={4} color="error">
+            <MailIcon />
+          </Badge>
+        </IconButton>
+        <p>Messages</p>
+      </MenuItem>
+      <MenuItem>
+        <IconButton
+          size="large"
+          aria-label="show 17 new notifications"
+          color="inherit"
+        >
+          <Badge badgeContent={17} color="error">
+            <NotificationsIcon />
+          </Badge>
+        </IconButton>
+        <p>Notifications</p>
+      </MenuItem>
+      <MenuItem onClick={handleProfileMenuOpen}>
+        <IconButton
+          size="large"
+          aria-label="account of current user"
+          aria-controls="primary-search-account-menu"
+          aria-haspopup="true"
+          color="inherit"
+        >
+          <AccountCircle />
+        </IconButton>
+        <p>Profile</p>
+      </MenuItem>
+    </Menu>
+  );
+
+  return (
+    <Box sx={{ flexGrow: 1 }}>
+      <AppBar sx={{ backgroundColor: '#a6b2ed' }} position="static">
+        <Toolbar>
+          {/* <IconButton
+            size="large"
+            edge="start"
+            color="inherit"
+            aria-label="open drawer"
+            sx={{ mr: 2 }}
+          >
+            <MenuIcon />
+          </IconButton> */}
+          <Box sx={{ padding: '10px 0px' }}>
+            <img className="radius" src={Logo} alt="logo" />
+          </Box>
+          <SearchBar></SearchBar>
+          <Box sx={{ flexGrow: 1 }} />
+          <Box sx={{ display: { xs: 'none', md: 'flex' } }}>
+            <Button sx={{ color: '#fff' }} component={Link} to="/">
+              Home
+            </Button>
+            <Button sx={{ color: '#fff' }} component={Link} to="/donacion">
+              Donacion
+            </Button>
+            <Button sx={{ color: '#fff' }} component={Link} to="/publicar">
+              Publicar
+            </Button>
+            <Button
+              sx={{ color: '#fff' }}
+              component={Link}
+              to="/misPublicaciones"
+            >
+              Mis Publicaciones
+            </Button>
+            <IconButton
+              size="large"
+              color="inherit"
+              component={Link}
+              to="/favoritos"
+            >
+              <FavoriteIcon />
+            </IconButton>
+            {/* <IconButton
+              size="large"
+              aria-label="show 17 new notifications"
+              color="inherit"
+            >
+              <Badge badgeContent={17} color="error">
+                <NotificationsIcon />
+              </Badge>
+            </IconButton> */}
+            <IconButton
+              size="large"
+              edge="end"
+              aria-label="account of current user"
+              aria-controls={menuId}
+              aria-haspopup="true"
+              onClick={handleProfileMenuOpen}
+              color="inherit"
+            >
+              <AccountCircle />
+            </IconButton>
+          </Box>
+          <Box sx={{ display: { xs: 'flex', md: 'none' } }}>
+            <IconButton
+              size="large"
+              aria-label="show more"
+              aria-controls={mobileMenuId}
+              aria-haspopup="true"
+              onClick={handleMobileMenuOpen}
+              color="inherit"
+            >
+              <MoreIcon />
+            </IconButton>
+          </Box>
+        </Toolbar>
+      </AppBar>
+      {renderMobileMenu}
+      {renderMenu}
+    </Box>
+  );
+}
